feat: fixing init so Nx can initialize the plugin properly after using `nx add`
