+++ packages/nx-electron-vite/src/generators/init/schema.json
@@ -1,35 +1,24 @@
 {
   "$schema": "https://json-schema.org/schema",
   "$id": "Init",
-  "title": "",
+  "title": "Init @erickrodrcodes/nx-electron-vite",
   "type": "object",
   "properties": {
-    "hostProject": {
-      "type": "string",
-      "alias": "hp",
-      "description": "Name of the host project",
-      "x-prompt": "Name of the host project",
-      "x-dropdown": "projects",
-      "x-priority": "important"
+    "skipPackageJson": {
+      "description": "Do not add dependencies to `package.json`.",
+      "type": "boolean",
+      "default": false
     },
-    "name": {
-      "type": "string",
-      "alias": "n",
-      "description": " you will give to your Electron Application",
-      "x-prompt": "Human readable name for your Electron Application?"
+    "skipFormat": {
+      "description": "Skip formatting files.",
+      "type": "boolean",
+      "default": false
     },
-    "author": {
-      "type": "string",
-      "alias": "a",
-      "description": "Author of the electron application",
-      "x-prompt": "Author of the electron application"
-    },
-    "description": {
-      "type": "string",
-      "alias": "d",
-      "description": "description the electron application",
-      "x-prompt": "description the electron application"
+    "installPluginDependencies": {
+      "description": "Installing Plugin dependencies. If you set this to false you might need to run your package manager install command to complete the install.",
+      "type": "boolean",
+      "default": true
     }
   },
-  "required": ["name", "hostProject", "author", "description"]
+  "required": []
 }
