+++ packages/nx-electron-vite/src/generators/setup-project/files/electron-nx-vite.config.ts.template
