+++ packages/nx-electron-vite/src/generators/setup-project/files/electron/dev-app-update.yml.template
