+++ packages/nx-electron-vite/src/generators/setup-project/files/electron/electron-builder.yml.template
