+++ packages/nx-electron-vite/src/generators/setup-project/files/electron/install.nsh.template
