+++ packages/nx-electron-vite/src/generators/setup-project/files/electron/main/electron-env.d.ts.template
