+++ packages/nx-electron-vite/src/generators/setup-project/files/electron/main/main.ts.template
