+++ packages/nx-electron-vite/src/generators/setup-project/files/electron/preload/preload.ts.template
