+++ packages/nx-electron-vite/src/generators/setup-project/files/electron/resources/certificate/Self.pfx
