+++ packages/nx-electron-vite/src/generators/setup-project/files/electron/resources/entitlements/macos/entitlements.mac.plist.template
