+++ packages/nx-electron-vite/src/generators/setup-project/files/electron/resources/icon/source/setup.png
