+++ packages/nx-electron-vite/src/generators/setup-project/files/package.json.template
