+++ packages/nx-electron-vite/src/generators/setup-project/mocks/jest.config.mock
