+++ packages/nx-electron-vite/src/generators/setup-project/mocks/project.json.mock
