+++ packages/nx-electron-vite/src/generators/setup-project/mocks/vite.config.implemented.mock
