+++ packages/nx-electron-vite/src/util/versions.ts
@@ -1,0 +1,6 @@
+export const electronBuilderVersion = '^24.13.3';
+export const electronVersion = '^30.0.2';
+export const vitePluginElectronRendererVersion = '^0.14.5';
+export const vitePluginElectronVersion = '^0.28.6';
+export const png2iconsVersion = '^2.0.1';
+export const suggestedNodeVersion = 'v18.18.0';
