+++ packages/nx-electron-vite/tsconfig.json
@@ -3,6 +3,7 @@
   "compilerOptions": {
     "module": "commonjs"
   },
+
   "files": [],
   "include": [],
   "references": [
